requirements.txt
pyserial==3.5
numpy==2.3.4
pandas==2.3.3
Pillow==12.0.0
customtkinter==5.2.2